Version 1.3.5
In this patch the installer has been improved and it detects whether a database does exist or not.
Besides, we are concealing an error caused by lack of browsercap.ini file. We strongly believe STORMY can be trustworthy, open and free only if we use our independent code. We inform that the statistics (browser name, operating system name) won't be available untill we solve this temporary problem. Thanks for understanding!