Version 1.5.8
In this update, we added the comments's smart filter. It uses profanity filter that directs the comments containing vulgar word and phrases to the moderation. Non-vulgar words are published instantly after adding. The smart filter recogines only polish words but it can be modified to decect any type of banned words (according to administrator's will).